backend/requirements.txt
click==8.1.3
colorama==0.4.5
ecdsa==0.18.0
Flask==2.1.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pyasn1==0.4.8
pycryptodome==3.15.0
python-dotenv==0.20.0
python-jose==3.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
wrapt==1.11.1
